Add cleared_cache fixture to clear cache before running test + some refractoring
- remove '--runslow' adaptor
- fixtures.py, just use conftests.py
- add `cleared_cache` autouse fixture, that also clears up cache before running any test
